requirements.txt
streamlit
fastapi
uvicorn
requests
beautifulsoup4
sentence-transformers
faiss-cpu